refactor olap_scan_prepare.h
